Created PostLoadable Module For Before Action Logic
